backend/requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
asgiref==3.8.1; python_version >= '3.8'
django==5.1.1; python_version >= '3.10'
django-binary-database-files==1.0.18; python_version < '3.13' and python_version >= '3.6'
django-cors-headers==4.4.0; python_version >= '3.8'
django-ninja==1.3.0; python_version >= '3.7'
fire==0.6.0
markdown-it-py==3.0.0; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
pydantic==2.8.2; python_version >= '3.8'
pydantic-core==2.20.1; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
rav==0.0.9; python_version >= '3.7'
rich==13.8.0; python_full_version >= '3.7.0'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.5.1; python_version >= '3.8'
termcolor==2.4.0; python_version >= '3.8'
typing-extensions==4.12.2; python_version < '3.13'
tzdata==2024.1; sys_platform == 'win32'
